Feature/#92 (찜한 정보 가져올때 캠핑장 정보를 조회해서 반환)

+++ src/main/java/com/commit/campus/repository/CampingRepository.java
@@ -9,6 +9,8 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface CampingRepository extends JpaRepository<Camping, Integer> {
+public interface CampingRepository extends JpaRepository<Camping, Long> {
+
     Page<Camping> findByCampIdIn(List<Long> reviewedCampIds, Pageable pageable);
+
 }
